+++ src/main.cxx
@@ -65,7 +65,7 @@ int main(int argc, char* argv[])
 	SPEED_TEST* t = new SPEED_TEST();
 	t->SM = sm;
 
-	uint32_t loops = 0;
+	uint64_t loops = 0;
 
 	if(stop){
 		loops = 1000000;
